Fix rustfmt issue
It doesn't like 'mod' statements with missing files, but is fine with missing 'use' statements
