+++ src/rust/engine/process_execution/bazel_protos/.gitignore
@@ -1,2 +1,3 @@
 src/*.rs
 !src/lib.rs
+!src/verification.rs
